Move sort logic outside of SortComponent
