+++ src/models/CompanyDecorator.ts
