+++ src/types/CompanyData.ts
